Adapt crm demo for showcase
Also change ConsileService to expect a mutable self reference
because it changes the console and it's more clear semantically
