+++ examples/showcase/Cargo.toml
@@ -6,3 +6,4 @@ authors = ["Denis Kolodin <deniskolodin@gmail.com>"]
 [dependencies]
 yew = { path = "../.." }
 counter = { path = "sub/counter" }
+crm = { path = "sub/crm" }
