+++ examples/showcase/sub/crm/Cargo.toml
@@ -6,4 +6,4 @@ authors = ["Denis Kolodin <deniskolodin@gmail.com>"]
 [dependencies]
 serde = "1"
 serde_derive = "1"
-yew = { path = "../.." }
+yew = { path = "../../../.." }
